move functions to config.py
