+++ vllm/model_executor/model_loader/tensorizer.py
@@ -13,13 +13,12 @@ from torch import nn
 from transformers import PretrainedConfig
 
 import vllm.envs as envs
-from vllm.config import ModelConfig, ParallelConfig
+from vllm.config import ModelConfig, ParallelConfig, set_current_vllm_config
 from vllm.engine.arg_utils import EngineArgs
 from vllm.engine.llm_engine import LLMEngine
 from vllm.logger import init_logger
 from vllm.model_executor.layers.vocab_parallel_embedding import (
     VocabParallelEmbedding)
-from vllm.plugins import set_current_vllm_config
 from vllm.utils import FlexibleArgumentParser
 
 tensorizer_error_msg = None
